--- a/.rubocop.yml
+++ b/.rubocop.yml
@@ -15,3 +15,6 @@ Style/DocumentDynamicEvalDefinition:
 
 Style/MultipleComparison:
   Enabled: false
+
+Style/NumericPredicate:
+  EnforcedStyle: comparison