@ms-ati 2024-09-30
Hi @tagliala! First I want to thank you for what appears to be an awareness campaign to help the community.

Can you explain why we would make this change here, in these two lines of code? One is in benchmark set up, the other is in one-time instrumentation of an object with MemoWise.

In neither case, is it yet clear to me that our trade-off would be for performance at the cost of (arguable) readability. Is that trade-off clear to you here? If so, can you help explain and make it clear?