@tagliala 2024-09-30
Hi,

> the other is in one-time instrumentation of an object with MemoWise.

In the specific case of memo wise, it is not going to provide any visible improvement, because as you mentioned it is a one time initialization. This is also stated in the opening post: "_This will not lead to performance improvements for the current implementation of memo_wise, I'm rather submitting this as a "reminder_"

> Is that trade-off clear to you here? If so, can you help explain and make it clear?

It is more than clear. The fact is that this cop is globally configured for readability, so it may lead to an usage of numeric predicates in contexts where they are not optimal, like loops. Again, this is not the case. I see `memo_wise` as a gem oriented to performance, so I proposed this change here in specific, but as a reminder and also to read other thoughts