@JacobEvelyn 2024-10-09
Thanks for making this PR, @tagliala! I appreciate the thoughtful commit message and interest in making future gem changes more performant.

Personally, I'm indifferent to this change. I generally like having performant defaults, but I also struggle to see a world in which we would change this gem to include `.zero?` or `.one?` in a hot path _at all_, let alone have it slip through code review _and_ our benchmarks. I'm not opposed to merging this however, since you took the time to make it.

@ms-ati do you have a preference?