@tagliala 2024-10-09
Welcome

>  I'm not opposed to merging this however, since you took the time to make it.

Thanks, and don't worry about the time.

I was also proposing to understand if that method could have made the difference in some circumstances, like large codebases, or if the new default may positively have impact on future development, but as far as I understand this is not a concern for memo_wise. And if it will ever be, it will be possible to make the change in the future