@JacobEvelyn 2024-10-09
> I was also proposing to understand if that method could have made the difference in some circumstances, like large codebases

On this point specifically I think it won't make a noticeable difference. That `.zero?` will typically run at load time (when `memo_wise` is called), and based on your benchmark it seems like both `.zero?` and `== 0` can execute many billions of times per second. I can't see a codebase ever having close to that many `memo_wise` calls so I don't think this will matter.